Happy Token BSC I64P98W3x6HapTok
Happy Token  founder have passed the KYC verification process.

Symbol: HAPPY

Contract: 0xC3cCA60Bdf643eCA0dC1e4943A41296ffBC5D4c8

Network: BSC

Telegram : https://t.me/HappyAnnouncement/2

Website: http://happytoken.site

WHAT IS THE HAPPY ECOSYSTEM?

Our ecosystem will be developed differently from many existing projects.
Our focus is community and we want to bring innovation into this world.

How will our ecosystem work?
A main token will be made that will be the core of the project.
It will be the MAIN TOKEN, which is the token where the whole ecosystem will feed.

How will the ecosystem be fed?
It will be fed by the creation of other tokens that will give the rewards in the main token.
This will allow the organic growth of the community with long term vision.
This will avoid the classic "When marketing?" "When calls?" "When? When? When?"

The ecosystem will be self-sustaining with the creation of secondary tokens for support and marketing focused on getting the word out about the ecosystem.